application/autocompletion/pyproject.toml
[tool.poetry]
name = "poi_autocompletion"
version = "0.1.0"
description = "autocompletion using elasticsearch and flask"
authors = ["joetsai"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
Flask-RESTful = "dotenv"
python-dotenv = "^0.19.0"
requests = "^2.26.0"
elasticsearch = "^7.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
